move analysis out of object model
