+++ generator/model.go
@@ -25,6 +25,7 @@ import (
 	"strconv"
 	"strings"
 
+	"github.com/go-swagger/go-swagger/analysis"
 	"github.com/go-swagger/go-swagger/spec"
 	"github.com/go-swagger/go-swagger/swag"
 )
@@ -153,8 +154,9 @@ func makeGenDefinitionHierarchy(name, pkg, container string, schema spec.Schema,
 	receiver := "m"
 	resolver := newTypeResolver("", specDoc)
 	resolver.ModelName = name
+	analyzed := analysis.New(specDoc.Spec())
 
-	di := discriminatorInfo(specDoc)
+	di := discriminatorInfo(analyzed)
 
 	pg := schemaGenContext{
 		Path:           "",
