+++ generator/operation.go
@@ -23,6 +23,7 @@ import (
 	"path/filepath"
 	"sort"
 
+	"github.com/go-swagger/go-swagger/analysis"
 	"github.com/go-swagger/go-swagger/httpkit"
 	"github.com/go-swagger/go-swagger/spec"
 	"github.com/go-swagger/go-swagger/swag"
@@ -46,8 +47,9 @@ func GenerateServerOperation(operationNames, tags []string, includeHandler, incl
 	if err != nil {
 		return err
 	}
+	analyzed := analysis.New(specDoc.Spec())
 
-	ops := gatherOperations(specDoc, operationNames)
+	ops := gatherOperations(analyzed, operationNames)
 
 	for operationName, opRef := range ops {
 		method, path, operation := opRef.Method, opRef.Path, opRef.Op
@@ -75,7 +77,7 @@ func GenerateServerOperation(operationNames, tags []string, includeHandler, incl
 			ClientPackage:        mangleName(swag.ToFileName(opts.ClientPackage), "client"),
 			ServerPackage:        serverPackage,
 			Operation:            *operation,
-			SecurityRequirements: specDoc.SecurityRequirementsFor(operation),
+			SecurityRequirements: analyzed.SecurityRequirementsFor(operation),
 			Principal:            opts.Principal,
 			Target:               filepath.Join(opts.Target, serverPackage),
 			Base:                 opts.Target,
@@ -88,6 +90,7 @@ func GenerateServerOperation(operationNames, tags []string, includeHandler, incl
 			DefaultProduces:      defaultProduces,
 			DefaultConsumes:      defaultConsumes,
 			Doc:                  specDoc,
+			Analyzed:             analyzed,
 		}
 		if err := generator.Generate(); err != nil {
 			return err
@@ -106,7 +109,7 @@ type operationGenerator struct {
 	ServerPackage        string
 	ClientPackage        string
 	Operation            spec.Operation
-	SecurityRequirements []spec.SecurityRequirement
+	SecurityRequirements []analysis.SecurityRequirement
 	Principal            string
 	Target               string
 	Base                 string
@@ -122,6 +125,7 @@ type operationGenerator struct {
 	DefaultProduces      string
 	DefaultConsumes      string
 	Doc                  *spec.Document
+	Analyzed             *analysis.Spec
 	WithContext          bool
 }
 
@@ -142,6 +146,7 @@ func (o *operationGenerator) Generate() error {
 	bldr.Operation = o.Operation
 	bldr.Authed = authed
 	bldr.Doc = o.Doc
+	bldr.Analyzed = o.Analyzed
 	bldr.DefaultScheme = o.DefaultScheme
 	bldr.DefaultProduces = o.DefaultProduces
 	bldr.DefaultImports = []string{filepath.ToSlash(filepath.Join(baseImport(o.Base), o.ModelsPackage))}
@@ -195,6 +200,7 @@ func (o *operationGenerator) Generate() error {
 		og.pkg = op.Package
 		og.cname = swag.ToGoName(op.Name)
 		og.Doc = o.Doc
+		og.Analyzed = o.Analyzed
 		og.Target = o.Target
 		og.APIPackage = o.APIPackage
 		og.WithContext = o.WithContext
@@ -212,6 +218,7 @@ type opGen struct {
 	IncludeParameters bool
 	IncludeResponses  bool
 	Doc               *spec.Document
+	Analyzed          *analysis.Spec
 	Target            string
 	APIPackage        string
 	WithContext       bool
@@ -226,7 +233,7 @@ func (o *opGen) Generate() error {
 		log.Println("generated handler", o.data.Package+"."+o.cname)
 	}
 
-	opParams := o.Doc.ParamsFor(o.data.Method, o.data.Path)
+	opParams := o.Analyzed.ParamsFor(o.data.Method, o.data.Path)
 	if o.IncludeParameters && len(opParams) > 0 {
 		if err := o.generateParameterModel(); err != nil {
 			return fmt.Errorf("parameters: %s", err)
@@ -304,6 +311,7 @@ type codeGenOpBuilder struct {
 	WithContext     bool
 	Operation       spec.Operation
 	Doc             *spec.Document
+	Analyzed        *analysis.Spec
 	Authed          bool
 	DefaultImports  []string
 	DefaultScheme   string
@@ -323,7 +331,7 @@ func (b *codeGenOpBuilder) MakeOperation() (GenOperation, error) {
 	operation := b.Operation
 	var params, qp, pp, hp, fp GenParameters
 	var hasQueryParams, hasFormParams, hasFileParams bool
-	for _, p := range b.Doc.ParamsFor(b.Method, b.Path) {
+	for _, p := range b.Analyzed.ParamsFor(b.Method, b.Path) {
 		cp, err := b.MakeParameter(receiver, resolver, p)
 		if err != nil {
 			return GenOperation{}, err
