+++ generator/parameter_test.go
@@ -31,7 +31,7 @@ func TestBodyParams(t *testing.T) {
 		t.FailNow()
 	}
 
-	_, _, op, ok := b.Doc.OperationForName("updateTask")
+	_, _, op, ok := b.Analyzed.OperationForName("updateTask")
 	if assert.True(t, ok) && assert.NotNil(t, op) {
 		resolver := &typeResolver{ModelsPackage: b.ModelsPackage, Doc: b.Doc}
 		resolver.KnownDefs = make(map[string]struct{})
@@ -59,7 +59,7 @@ func TestBodyParams(t *testing.T) {
 		t.FailNow()
 	}
 
-	_, _, op, ok = b.Doc.OperationForName("createTask")
+	_, _, op, ok = b.Analyzed.OperationForName("createTask")
 	if assert.True(t, ok) && assert.NotNil(t, op) {
 		resolver := &typeResolver{ModelsPackage: b.ModelsPackage, Doc: b.Doc}
 		resolver.KnownDefs = make(map[string]struct{})
@@ -261,7 +261,7 @@ type paramTestContext struct {
 }
 
 func (ctx *paramTestContext) assertParameter(t testing.TB) bool {
-	_, _, op, err := ctx.B.Doc.OperationForName(ctx.OpID)
+	_, _, op, err := ctx.B.Analyzed.OperationForName(ctx.OpID)
 	if assert.True(t, err) && assert.NotNil(t, op) {
 		resolver := &typeResolver{ModelsPackage: ctx.B.ModelsPackage, Doc: ctx.B.Doc}
 		resolver.KnownDefs = make(map[string]struct{})
