+++ httpkit/middleware/context.go
@@ -18,6 +18,7 @@ import (
 	"net/http"
 	"strings"
 
+	"github.com/go-swagger/go-swagger/analysis"
 	"github.com/go-swagger/go-swagger/errors"
 	"github.com/go-swagger/go-swagger/httpkit"
 	"github.com/go-swagger/go-swagger/httpkit/middleware/untyped"
@@ -55,10 +56,11 @@ func (fn ResponderFunc) WriteResponse(rw http.ResponseWriter, pr httpkit.Produce
 // Context is a type safe wrapper around an untyped request context
 // used throughout to store request context with the gorilla context module
 type Context struct {
-	spec    *spec.Document
-	api     RoutableAPI
-	router  Router
-	formats strfmt.Registry
+	spec     *spec.Document
+	analyzer *analysis.Spec
+	api      RoutableAPI
+	router   Router
+	formats  strfmt.Registry
 }
 
 type routableUntypedAPI struct {
@@ -73,10 +75,11 @@ func newRoutableUntypedAPI(spec *spec.Document, api *untyped.API, context *Conte
 	if spec == nil || api == nil {
 		return nil
 	}
-	for method, hls := range spec.Operations() {
+	analyzer := analysis.New(spec.Spec())
+	for method, hls := range analyzer.Operations() {
 		um := strings.ToUpper(method)
 		for path, op := range hls {
-			schemes := spec.SecurityDefinitionsFor(op)
+			schemes := analyzer.SecurityDefinitionsFor(op)
 
 			if oh, ok := api.OperationHandlerFor(method, path); ok {
 				if handlers == nil {
@@ -158,13 +161,21 @@ func (r *routableUntypedAPI) DefaultConsumes() string {
 
 // NewRoutableContext creates a new context for a routable API
 func NewRoutableContext(spec *spec.Document, routableAPI RoutableAPI, routes Router) *Context {
-	ctx := &Context{spec: spec, api: routableAPI}
+	var an *analysis.Spec
+	if spec != nil {
+		an = analysis.New(spec.Spec())
+	}
+	ctx := &Context{spec: spec, api: routableAPI, analyzer: an}
 	return ctx
 }
 
 // NewContext creates a new context wrapper
 func NewContext(spec *spec.Document, api *untyped.API, routes Router) *Context {
-	ctx := &Context{spec: spec}
+	var an *analysis.Spec
+	if spec != nil {
+		an = analysis.New(spec.Spec())
+	}
+	ctx := &Context{spec: spec, analyzer: an}
 	ctx.api = newRoutableUntypedAPI(spec, api, ctx)
 	return ctx
 }
@@ -207,7 +218,7 @@ func (c *Context) BasePath() string {
 
 // RequiredProduces returns the accepted content types for responses
 func (c *Context) RequiredProduces() []string {
-	return c.spec.RequiredProduces()
+	return c.analyzer.RequiredProduces()
 }
 
 // BindValidRequest binds a params object to a request but only when the request is valid
