+++ httpkit/middleware/router.go
@@ -19,6 +19,7 @@ import (
 	"regexp"
 	"strings"
 
+	"github.com/go-swagger/go-swagger/analysis"
 	"github.com/go-swagger/go-swagger/errors"
 	"github.com/go-swagger/go-swagger/httpkit"
 	"github.com/go-swagger/go-swagger/spec"
@@ -89,11 +90,11 @@ func newRouter(ctx *Context, next http.Handler) http.Handler {
 		}
 		// Not found, check if it exists in the other methods first
 		if others := ctx.AllowedMethods(r); len(others) > 0 {
-			ctx.Respond(rw, r, ctx.spec.RequiredProduces(), nil, errors.MethodNotAllowed(r.Method, others))
+			ctx.Respond(rw, r, ctx.analyzer.RequiredProduces(), nil, errors.MethodNotAllowed(r.Method, others))
 			return
 		}
 
-		ctx.Respond(rw, r, ctx.spec.RequiredProduces(), nil, errors.NotFound("path %s was not found", r.URL.Path))
+		ctx.Respond(rw, r, ctx.analyzer.RequiredProduces(), nil, errors.NotFound("path %s was not found", r.URL.Path))
 	})
 }
 
@@ -117,9 +118,10 @@ type Router interface {
 }
 
 type defaultRouteBuilder struct {
-	spec    *spec.Document
-	api     RoutableAPI
-	records map[string][]denco.Record
+	spec     *spec.Document
+	analyzer *analysis.Spec
+	api      RoutableAPI
+	records  map[string][]denco.Record
 }
 
 type defaultRouter struct {
@@ -130,9 +132,10 @@ type defaultRouter struct {
 
 func newDefaultRouteBuilder(spec *spec.Document, api RoutableAPI) *defaultRouteBuilder {
 	return &defaultRouteBuilder{
-		spec:    spec,
-		api:     api,
-		records: make(map[string][]denco.Record),
+		spec:     spec,
+		analyzer: analysis.New(spec.Spec()),
+		api:      api,
+		records:  make(map[string][]denco.Record),
 	}
 }
 
@@ -140,7 +143,7 @@ func newDefaultRouteBuilder(spec *spec.Document, api RoutableAPI) *defaultRouteB
 func DefaultRouter(spec *spec.Document, api RoutableAPI) Router {
 	builder := newDefaultRouteBuilder(spec, api)
 	if spec != nil {
-		for method, paths := range spec.Operations() {
+		for method, paths := range builder.analyzer.Operations() {
 			for path, operation := range paths {
 				builder.AddRoute(method, path, operation)
 			}
@@ -207,10 +210,10 @@ func (d *defaultRouteBuilder) AddRoute(method, path string, operation *spec.Oper
 	mn := strings.ToUpper(method)
 
 	if handler, ok := d.api.HandlerFor(method, path); ok {
-		consumes := d.spec.ConsumesFor(operation)
-		produces := d.spec.ProducesFor(operation)
-		parameters := d.spec.ParamsFor(method, path)
-		definitions := d.spec.SecurityDefinitionsFor(operation)
+		consumes := d.analyzer.ConsumesFor(operation)
+		produces := d.analyzer.ProducesFor(operation)
+		parameters := d.analyzer.ParamsFor(method, path)
+		definitions := d.analyzer.SecurityDefinitionsFor(operation)
 
 		record := denco.NewRecord(pathConverter.ReplaceAllString(path, ":$1"), &routeEntry{
 			Operation:      operation,
